web-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.2.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
python-dateutil==2.8.2
email-validator==2.1.0